[#109] [Bug] 결과 공유 페이지에 navigator.share 버튼이 보이지 않는 현상
- kakao 공유 메시지 포맷을 커스텀이 아닌 메타 정보로 보이도록 sendDefault에서 sendScrap 함수로 변경
- kakaoSDK에 특화된 타입을 KakaoSDK.ts로 이동함
