+++ src/components/common/Modal.ts
@@ -24,7 +24,8 @@ export const Background = styled.div<{ show: boolean }>`
   width: 100%;
   height: 100%;
   z-index: 1;
-  background-color: rgba(0, 0, 0, 0.3);
+  backdrop-filter: blur(2px);
+  background-color: rgba(0, 0, 0, 0.5);
   animation: 0.2s ${props => (props.show ? FadeIn : FadeOut)};
 `;
